fix: more inline with source; update poetry.lock
