+++ pyproject.toml
@@ -13,7 +13,7 @@ keywords = ["terminal", "gui"]
 python = ">=3.7"
 blessed = "1.21.0"
 
-[tool.poetry.dev-dependencies]
+[tool.poetry.group.dev.dependencies]
 pytest = "^7.2.0"
 mypy = "^1.4"
 pre-commit = "^2.20.0"
