+++ src/pick/__init__.py
@@ -1,10 +1,24 @@
-import curses
-import textwrap
+#!/usr/bin/env python
+
 from collections import namedtuple
 from dataclasses import dataclass, field
-from typing import Any, Container, Generic, Iterable, List, Optional, Sequence, Tuple, TypeVar, Union
+from typing import (
+    Any,
+    Generic,
+    Iterable,
+    List,
+    Optional,
+    Sequence,
+    Tuple,
+    TypeVar,
+    Union,
+)
+import blessed
 
-__all__ = ["Picker", "pick", "Option"]
+__all__ = ["pick", "Picker", "Option"]
+
+SYMBOL_CIRCLE_FILLED = "(x)"
+SYMBOL_CIRCLE_EMPTY = "( )"
 
 
 @dataclass
@@ -14,35 +28,68 @@ class Option:
     description: Optional[str] = None
     enabled: bool = True
 
+    def __str__(self) -> str:
+        return (
+            f"{self.label}{' (' + self.description + ')' if self.description else ''}"
+        )
 
-KEYS_ENTER = (curses.KEY_ENTER, ord("\n"), ord("\r"))
-KEYS_UP = (curses.KEY_UP, ord("k"))
-KEYS_DOWN = (curses.KEY_DOWN, ord("j"))
-KEYS_SELECT = (curses.KEY_RIGHT, ord(" "))
 
-SYMBOL_CIRCLE_FILLED = "(x)"
-SYMBOL_CIRCLE_EMPTY = "( )"
-
-OPTION_T = TypeVar("OPTION_T", str, Option)
+OPTION_T = Union[Option, str]
 PICK_RETURN_T = Tuple[OPTION_T, int]
 
-Position = namedtuple('Position', ['y', 'x'])
+Position = namedtuple("Position", ["y", "x"])
+
+
+def _display_screen(
+    term: blessed.Terminal,
+    indicator: str,
+    title: Optional[str],
+    choices: Sequence[OPTION_T],
+    index: int,
+    selected_indexes: list[int],
+    multiselect: bool,
+) -> None:
+    if title:
+        print(title)
+
+    for idx, val in enumerate(choices):
+        selectable = ""
+        if isinstance(val, Option) and not val.enabled:
+            selectable = term.gray35
+
+        is_selected = ""
+        if multiselect:
+            is_selected = (
+                f"{SYMBOL_CIRCLE_EMPTY} "
+                if idx not in selected_indexes
+                else f"{SYMBOL_CIRCLE_FILLED} "
+            )
+
+        if idx == index:
+            print(f"{indicator} {selectable}{is_selected}{val}{term.normal}")
+        else:
+            print(
+                f"{' ' * (len(indicator) + 1)}{selectable}{is_selected}{val}{term.normal}"
+            )
+
 
 @dataclass
-class Picker(Generic[OPTION_T]):
+class Picker:
     options: Sequence[OPTION_T]
     title: Optional[str] = None
     indicator: str = "*"
     default_index: int = 0
     multiselect: bool = False
     min_selection_count: int = 0
-    selected_indexes: List[int] = field(init=False, default_factory=list)
-    index: int = field(init=False, default=0)
-    screen: Optional["curses._CursesWindow"] = None
+    selected_indexes: List[int] = field(init=True, default_factory=list)
+    index: int = field(init=True, default=0)
     position: Position = Position(0, 0)
     clear_screen: bool = True
-    quit_keys: Optional[Union[Container[int], Iterable[int]]] = None
+    quit_keys: Optional[Iterable[int]] = None
+    term: blessed.Terminal = None
+    # ) -> List[PICK_RETURN_T]:
 
+    # screen: Optional["curses._CursesWindow"] = None
     def __post_init__(self) -> None:
         if len(self.options) == 0:
             raise ValueError("options should not be an empty list")
@@ -55,7 +102,9 @@ class Picker(Generic[OPTION_T]):
                 "min_selection_count is bigger than the available options, you will not be able to make any selection"
             )
 
-        if all(isinstance(option, Option) and not option.enabled for option in self.options):
+        if all(
+            isinstance(option, Option) and not option.enabled for option in self.options
+        ):
             raise ValueError(
                 "all given options are disabled, you must at least have one enabled option."
             )
@@ -85,148 +134,89 @@ class Picker(Generic[OPTION_T]):
 
     def mark_index(self) -> None:
         if self.multiselect:
-            if self.index in self.selected_indexes:
-                self.selected_indexes.remove(self.index)
-            else:
-                self.selected_indexes.append(self.index)
+            item = self.options[self.index]
+            if (isinstance(item, Option) and item.enabled) or not isinstance(
+                item, Option
+            ):
+                if self.index in self.selected_indexes:
+                    self.selected_indexes.remove(self.index)
+                else:
+                    self.selected_indexes.append(self.index)
 
     def get_selected(self) -> Union[List[PICK_RETURN_T], PICK_RETURN_T]:
-        """return the current selected option as a tuple: (option, index)
+        """Return the current selected option as a tuple: (option, index)
         or as a list of tuples (in case multiselect==True)
         """
         if self.multiselect:
-            return_tuples = []
-            for selected in self.selected_indexes:
-                return_tuples.append((self.options[selected], selected))
-            return return_tuples
+            return (
+                [(self.options[idx], idx) for idx in self.selected_indexes]
+                if self.multiselect
+                else (self.options[self.index], self.index)
+            )
         else:
             return self.options[self.index], self.index
 
-    def get_title_lines(self, *, max_width: int = 80) -> List[str]:
-        if self.title:
-            return textwrap.fill(self.title, max_width - 2, drop_whitespace=False).split("\n") + [""]
-        return []
-
-    def get_option_lines(self) -> List[str]:
-        lines: List[str] = []
-        for index, option in enumerate(self.options):
-            if index == self.index:
-                prefix = self.indicator
-            else:
-                prefix = len(self.indicator) * " "
-
-            if self.multiselect:
-                symbol = (
-                    SYMBOL_CIRCLE_FILLED
-                    if index in self.selected_indexes
-                    else SYMBOL_CIRCLE_EMPTY
-                )
-                prefix = f"{prefix} {symbol}"
-
-            option_as_str = option.label if isinstance(option, Option) else option
-            lines.append(f"{prefix} {option_as_str}")
-
-        return lines
-
-    def get_lines(self, *, max_width: int = 80) -> Tuple[List[str], int]:
-        title_lines = self.get_title_lines(max_width=max_width)
-        option_lines = self.get_option_lines()
-        lines = title_lines + option_lines
-        current_line = self.index + len(title_lines) + 1
-        return lines, current_line
-
-    def draw(self, screen: "curses._CursesWindow") -> None:
-        """draw the curses ui on the screen, handle scroll if needed"""
-        if self.clear_screen:
-            screen.clear()
-
-        y, x = self.position  # start point
-
-        max_y, max_x = screen.getmaxyx()
-        max_rows = max_y - y  # the max rows we can draw
-
-        lines, current_line = self.get_lines(max_width=max_x)
-
-        # calculate how many lines we should scroll, relative to the top
-        scroll_top = 0
-        if current_line > max_rows:
-            scroll_top = current_line - max_rows
-
-        lines_to_draw = lines[scroll_top : scroll_top + max_rows]
-
-        description_present = False
-        for option in self.options:
-            if isinstance(option, Option) and option.description is not None:
-                description_present = True
-                break
-
-        title_length = len(self.get_title_lines(max_width=max_x))
-
-        for i, line in enumerate(lines_to_draw):
-            if description_present and i > title_length:
-                screen.addnstr(y, x, line, max_x // 2 - 2)
-            else:
-                screen.addnstr(y, x, line, max_x - 2)
-            y += 1
-
-        option = self.options[self.index]
-        if isinstance(option, Option) and option.description is not None:
-            description_lines = textwrap.fill(option.description, max_x // 2 - 2).split('\n')
-
-            for i, line in enumerate(description_lines):
-                screen.addnstr(i + title_length, max_x // 2, line, max_x - 2)
-
-        screen.refresh()
-
-    def run_loop(
-        self, screen: "curses._CursesWindow", position: Position
-    ) -> Union[List[PICK_RETURN_T], PICK_RETURN_T]:
-        while True:
-            self.draw(screen)
-            c = screen.getch()
-            if self.quit_keys is not None and c in self.quit_keys:
-                if self.multiselect:
-                    return []
-                else:
-                    return None, -1
-            elif c in KEYS_UP:
-                self.move_up()
-            elif c in KEYS_DOWN:
-                self.move_down()
-            elif c in KEYS_ENTER:
-                if (
-                    self.multiselect
-                    and len(self.selected_indexes) < self.min_selection_count
-                ):
-                    continue
-                return self.get_selected()
-            elif c in KEYS_SELECT and self.multiselect:
-                self.mark_index()
-
-    def config_curses(self) -> None:
-        try:
-            # use the default colors of the terminal
-            curses.use_default_colors()
-            # hide the cursor
-            curses.curs_set(0)
-        except Exception:
-            # Curses failed to initialize color support, eg. when TERM=vt100
-            curses.initscr()
-
-    def _start(self, screen: "curses._CursesWindow"):
-        self.config_curses()
-        return self.run_loop(screen, self.position)
-
     def start(self):
-        if self.screen:
-            # Given an existing screen
-            # don't make any lasting changes
-            last_cur = curses.curs_set(0)
-            ret = self.run_loop(self.screen, self.position)
-            if last_cur:
-                curses.curs_set(last_cur)
-            return ret
-        return curses.wrapper(self._start)
+        _quit_keys: list[str] = (
+            []
+            if self.quit_keys is None
+            else [chr(key_code) for key_code in self.quit_keys]
+        )
+        errmsg = ""
+
+        with self.term.fullscreen(), self.term.cbreak(), self.term.hidden_cursor():
+            print(self.term.clear())
+            _display_screen(
+                self.term,
+                self.indicator,
+                self.title,
+                self.options,
+                self.index,
+                self.selected_indexes,
+                self.multiselect,
+            )
+
+            selection_inprogress = True
+            while selection_inprogress:
+                key = self.term.inkey()
+                if key.is_sequence or key == " ":
+                    if key.name in {"KEY_TAB", "KEY_DOWN"}:
+                        self.move_down()
+                    elif key.name == "KEY_UP":
+                        self.move_up()
+                    elif key == " " and self.multiselect:
+                        self.mark_index()
+                    elif key.name == "KEY_ENTER":
+                        if (
+                            self.multiselect
+                            and len(self.selected_indexes) < self.min_selection_count
+                        ):
+                            errmsg = f"{self.term.red}Must select at least {self.min_selection_count} entry(s)!{self.term.normal}"
+                        else:
+                            return self.get_selected()
+                else:
+                    if key.lower() in _quit_keys:
+                        return None, -1
+
+                self.index = self.index % len(self.options)
+
+                print(self.term.clear())
+
+                if errmsg:
+                    print(errmsg)
+                    errmsg = ""
+
+                _display_screen(
+                    self.term,
+                    self.indicator,
+                    self.title,
+                    self.options,
+                    self.index,
+                    self.selected_indexes,
+                    self.multiselect,
+                )
+
+        return self.get_selected()
 
 
 def pick(
@@ -236,21 +226,70 @@ def pick(
     default_index: int = 0,
     multiselect: bool = False,
     min_selection_count: int = 0,
-    screen: Optional["curses._CursesWindow"] = None,
     position: Position = Position(0, 0),
     clear_screen: bool = True,
-    quit_keys: Optional[Union[Container[int], Iterable[int]]] = None,
-):
-    picker: Picker = Picker(
-        options,
-        title,
-        indicator,
-        default_index,
-        multiselect,
-        min_selection_count,
-        screen,
-        position,
-        clear_screen,
-        quit_keys,
-    )
-    return picker.start()
+    quit_keys: Optional[Iterable[int]] = None,
+) -> Union[List[PICK_RETURN_T], Optional[PICK_RETURN_T]]:
+    term = blessed.Terminal()
+    picked = None
+
+    with term.fullscreen(), term.cbreak():
+        print(Picker(["a"], "a"))
+        print("???")
+        picked = Picker(
+            options=options,
+            title=title,
+            indicator=indicator,
+            default_index=default_index,
+            multiselect=multiselect,
+            min_selection_count=min_selection_count,
+            selected_indexes=[],
+            index=0,
+            clear_screen=clear_screen,
+            position=position,
+            quit_keys=quit_keys,
+            term=term,
+        ).start()
+
+    return picked if multiselect else (picked[0] if picked else None)
+
+
+print(
+    "Picked: ",
+    pick(
+        [
+            Option(
+                "Option 1",
+                "option 1",
+                "this is option 1 and is not selectable",
+                enabled=False,
+            ),
+            "option 2",
+            "option 3",
+            Option(
+                "Option 4",
+                "option 4",
+                "this is option 4 and selectable",
+                enabled=True,
+            ),
+        ],
+        "(Up/down/tab to move; space to select/de-select; Enter to continue)",
+        indicator="=>",
+        multiselect=True,
+        quit_keys=[ord("q")],
+        clear_screen=False,
+        min_selection_count=2,
+    ),
+)
+print()
+
+print(
+    "Picked: ",
+    pick(
+        ["Choice1", "choice 2", "choice3"],
+        "(Up/down/tab to move; Enter to select)",
+        indicator="=>",
+        multiselect=False,
+        quit_keys=[ord("q")],
+    ),
+)
